pyproject.toml
[project]
name = "assistant"
version = "1.0.0"
description = "Personal AI Assistant v1"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.32.0",
    "pydantic>=2.9.0",
    "pydantic-settings>=2.6.0",
    "pydantic-ai[anthropic]>=0.0.14",
    "anthropic>=0.40.0",
    "pyyaml>=6.0.2",
    "python-multipart>=0.0.12",
    "httpx>=0.28.0",
    "structlog>=24.4.0",
    "python-dotenv>=1.0.1",
    "anyio>=4.7.0",
    "jinja2>=3.1.6",
    "itsdangerous>=2.2.0",
    "aiogram>=3.26.0",
    "telegramify-markdown>=1.0.0",
    "genai-prices>=0.0.55",
    "rank-bm25>=0.2.2",
    "tavily-python>=0.5.0",
    "mem0ai>=0.1.0",
]

[dependency-groups]
dev = [
    "ruff>=0.8.0",
    "mypy>=1.13.0",
    "pytest>=8.3.0",
    "pytest-asyncio>=0.24.0",
    "pytest-cov>=6.0.0",
    "types-pyyaml>=6.0.12",
]

[tool.uv]
default-groups = ["dev"]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/assistant"]

[tool.ruff]
line-length = 100
target-version = "py312"
src = ["src"]

[tool.ruff.lint]
select = ["E", "F", "I", "UP", "B", "SIM", "ANN"]
ignore = ["ANN101", "ANN102", "ANN401", "B008"]

[tool.ruff.lint.isort]
known-first-party = ["assistant"]

[tool.mypy]
python_version = "3.12"
strict = true
ignore_missing_imports = true
explicit_package_bases = true
mypy_path = "src"

[tool.pytest.ini_options]
testpaths = ["tests"]
asyncio_mode = "auto"
